e2e fixes for NG
