+++ e2e/vpc/vpc_create_volume_attach_detach_volume.go
@@ -101,11 +101,3 @@ var _ = Describe("ibmcloud-storage-volume-lib", func() {
 		fmt.Printf("\n\n")
 	})
 })
-
-func getenv(key, fallback string) string {
-	value := os.Getenv(key)
-	if len(value) == 0 {
-		return fallback
-	}
-	return value + "-"
-}
